+++ include/utils/pixel.h
@@ -17,6 +17,7 @@ void argb2rgba32(unsigned char* image, unsigned int w, unsigned int h);
 
 // Returns newly allocated buffer
 unsigned char* argb16to32(unsigned char* image, unsigned int w, unsigned int h);
+unsigned char* grayscale2rgba(unsigned char* image, unsigned int w, unsigned int h);
 
 unsigned char* scaleBuffer(unsigned char* image, unsigned int* w, unsigned int* h,
                            unsigned int bpp);